refactoring code, add `CanActivateChild` to guards and update readme docs
