+++ projects/ngx-jwt-auth/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@dekh/ngx-jwt-auth",
-	"version": "1.0.2",
+	"version": "1.0.3",
 	"license": "MIT",
 	"repository": {
 		"type": "GitHub",
