+++ projects/ngx-jwt-auth/src/lib/const/index.ts
@@ -1,2 +1,3 @@
 export * from './jwt-auth-token-storage-key';
+export * from './predefined-token-storages';
 export * from './token-storage-key';
