+++ projects/ngx-jwt-auth/src/lib/enums/auth-token.enum.ts
