+++ projects/ngx-jwt-auth/src/lib/enums/samesite-option.enum.ts
