+++ projects/ngx-jwt-auth/src/lib/token-storages/index.ts
@@ -2,5 +2,5 @@ export * from './base-token-storage';
 export * from './cookies-token-storage';
 export * from './in-memory-token-storage';
 export * from './local-storage-token-storage';
-export * from './predefined-token-storages';
 export * from './session-storage-token-storage';
+
